+++ src/wkt_conv.cpp
@@ -220,8 +220,17 @@ bool srs_is_projected(std::string srs) {
 //' the same system. This is a wrapper for `OSRIsSame()` in the GDAL Spatial 
 //' Reference System C API.
 //'
-//' @param srs1 Character OGC WKT string for a spatial reference system
-//' @param srs2 Character OGC WKT string for a spatial reference system
+//' @param srs1 Character string. OGC WKT for a spatial reference system.
+//' @param srs2 Character string. OGC WKT for a spatial reference system.
+//' @param criterion Character string. One of `STRICT`, `EQUIVALENT`,
+//' `EQUIVALENT_EXCEPT_AXIS_ORDER_GEOGCRS`.
+//' Defaults to `EQUIVALENT_EXCEPT_AXIS_ORDER_GEOGCRS`.
+//' @param ignore_axis_mapping Logical scalar. If `TRUE`, sets
+//' `IGNORE_DATA_AXIS_TO_SRS_AXIS_MAPPING=YES` in the call to `OSRIsSameEx()`
+//' in the GDAL Spatial Reference System API. Defaults to `NO`.
+//' @param ignore_coord_epoch Logical scalar. If `TRUE`, sets
+//' `IGNORE_COORDINATE_EPOCH=YES` in the call to `OSRIsSameEx()`
+//' in the GDAL Spatial Reference System API. Defaults to `NO`.
 //' @return Logical. `TRUE` if these two spatial references describe the same 
 //' system, otherwise `FALSE`.
 //'
@@ -235,7 +244,10 @@ bool srs_is_projected(std::string srs) {
 //' srs_is_same(ds$getProjectionRef(), epsg_to_wkt(5070))
 //' ds$close()
 // [[Rcpp::export]]
-bool srs_is_same(std::string srs1, std::string srs2) {
+bool srs_is_same(std::string srs1, std::string srs2,
+		std::string criterion = "",
+		bool ignore_axis_mapping = false,
+		bool ignore_coord_epoch = false) {
 
 	OGRSpatialReferenceH hSRS1 = OSRNewSpatialReference(NULL);
 	OGRSpatialReferenceH hSRS2 = OSRNewSpatialReference(NULL);
@@ -250,7 +262,34 @@ bool srs_is_same(std::string srs1, std::string srs2) {
 	if (OSRImportFromWkt(hSRS2, &pszWKT2) != OGRERR_NONE)
 		Rcpp::stop("Error importing SRS from user input.");
 	
-	return OSRIsSame(hSRS1, hSRS2);
+	std::vector<char *> opt_list;
+	std::string str_axis;
+	std::string str_epoch;
+	
+	if (criterion != "") {
+		criterion = "CRITERION=" + criterion;
+		opt_list.push_back((char *) criterion.c_str());
+	}
+	
+	if (ignore_axis_mapping) {
+		str_axis = "IGNORE_DATA_AXIS_TO_SRS_AXIS_MAPPING=YES";
+	}
+	else {
+		str_axis = "IGNORE_DATA_AXIS_TO_SRS_AXIS_MAPPING=NO";
+	}
+	opt_list.push_back((char *) str_axis.c_str());
+	
+	if (ignore_coord_epoch) {
+		str_epoch = "IGNORE_COORDINATE_EPOCH=YES";
+	}
+	else {
+		str_epoch = "IGNORE_COORDINATE_EPOCH=NO";
+	}
+	opt_list.push_back((char *) str_epoch.c_str());
+	
+	opt_list.push_back(NULL);
+	
+	return OSRIsSameEx(hSRS1, hSRS2, opt_list.data());
 }
 
 //' Get the bounding box of a geometry specified in OGC WKT format